Split of CSS, JS and images from HTML-Controller, added reporting page (still needs token though)

+++ src/main/java/doko/database/token/Token.java
@@ -14,35 +14,35 @@ import javax.persistence.Table;
 public class Token {
 
 	@Id
-    @GeneratedValue(strategy = GenerationType.AUTO)
-    private long id;
-	
-	@Column(name="user_id", nullable = false)
-    private Integer userId;
-	
-	@Column(name="token_value", nullable = false)
-    private String tokenValue;
-	
-	@Column(name="expiration_date", nullable = false)
-    private Date expirationDate;
-	
-	public Token(Integer userId, String tokenValue, Date expirationDate) {
+	@GeneratedValue(strategy = GenerationType.AUTO)
+	private long id;
+
+	@Column(name = "user_id", nullable = false)
+	private Long userId;
+
+	@Column(name = "token_value", nullable = false)
+	private String tokenValue;
+
+	@Column(name = "expiration_date", nullable = false)
+	private Date expirationDate;
+
+	public Token(Long userId, String tokenValue, Date expirationDate) {
 		this.userId = userId;
 		this.tokenValue = tokenValue;
 		this.expirationDate = expirationDate;
 	}
-	
+
 	public Token() {
 	}
-	
-	public Integer getUserId() {
+
+	public Long getUserId() {
 		return userId;
 	}
-	
+
 	public String getTokenValue() {
 		return tokenValue;
 	}
-	
+
 	public Date getExpirationDate() {
 		return expirationDate;
 	}
